add init philos

+++ philo/include/philo.h
@@ -6,7 +6,7 @@
 /*   By: kfukuhar <kfukuhar@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/23 14:52:20 by kfukuhar          #+#    #+#             */
-/*   Updated: 2024/09/23 19:08:56 by kfukuhar         ###   ########.fr       */
+/*   Updated: 2024/09/23 19:21:37 by kfukuhar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,10 +40,10 @@ typedef struct s_fork
 
 typedef struct s_philo
 {
-	int					id;
-	long int			meals_counter;
+	size_t				id;
+	size_t				meals_counter;
 	bool				full;
-	long int			last_meal_time;
+	size_t				last_meal_time;
 	t_fork				*l_fork;
 	t_fork				*r_fork;
 	pthread_t			thread_id;
